+++ react-components/src/components/MarkLessonAsComplete/MarkLessonAsComplete.jsx
@@ -46,7 +46,7 @@ const MarkLessonAsComplete = ({ endpoint }) => {
   return (
     <div className="mark-lesson-as-complete">
       <legend>
-        <h2 className="h-m text-white p-b-s">Lesson complete?</h2>
+        <h2 className="h-m text-white p-b-s" aria-hidden="true">Lesson complete?</h2>
       </legend>
       <div className="great-checkbox great-checkbox--large">
         <input
@@ -58,8 +58,7 @@ const MarkLessonAsComplete = ({ endpoint }) => {
           onClick={markCompleted}
           checked={Boolean(isComplete)}
         />
-        <label htmlFor="markascomplete_checkbox">{labelText}</label>
-        <div className="visually-hidden" role="status">{isComplete && isChecked && 'Great! Progress saved'}</div>
+        <label htmlFor="markascomplete_checkbox"><span className="visually-hidden">Lesson complete?</span><span aria-hidden="true">{labelText}</span></label>
       </div>
     </div>
   )