+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Pre-release changes - please put everything in the appropriate category below
 
+### Enhancements
+
+### Fixed bugs
+- GP2-2938 - DAC_Nondescriptive_Labels_02 Lesson complete checkbox
+
 ## [1.11.0](https://github.com/uktrade/great-cms/releases/tag/1.11.0)
 [Full Changelog](https://github.com/uktrade/great-cms/compare/1.10.0...1.11.0)